integration_tests/python_tests/requirements.txt
pytest>=7.0.0
snowflake-ml-python>=1.0.0
snowflake-snowpark-python>=1.0.0


